verify-replciation not tested
